# replitui/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Attendance System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AttendanceRecordStatus =
  (typeof AttendanceRecordStatus)[keyof typeof AttendanceRecordStatus];

export const AttendanceRecordStatus = {
  present: "present",
  duplicate: "duplicate",
  invalid: "invalid",
} as const;

export interface AttendanceRecord {
  id: number;
  name: string;
  uid: string;
  status: AttendanceRecordStatus;
  lectureNumber: number;
  timestamp: string;
}

export interface CreateAttendanceBody {
  name: string;
  uid: string;
  lectureNumber: number;
}

export interface StudentSummary {
  name: string;
  uid: string;
  totalScans: number;
  presentCount: number;
  duplicateCount: number;
  invalidCount: number;
}

export interface LectureStat {
  lectureNumber: number;
  totalScans: number;
  presentCount: number;
  duplicateCount: number;
  invalidCount: number;
}

export interface ActiveLecture {
  /** @nullable */
  lectureNumber: number | null;
  isActive: boolean;
  /** @nullable */
  startedAt: string | null;
}

export interface StartLectureBody {
  lectureNumber: number;
}

export interface ResetResponse {
  success: boolean;
  message: string;
}

export type GetAttendanceParams = {
  lectureNumber?: number;
};