# replitui/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Attendance System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all attendance records, optionally filtered by lecture
 * @summary Get all attendance records
 */
export const GetAttendanceQueryParams = zod.object({
  lectureNumber: zod.coerce.number().optional(),
});

export const GetAttendanceResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  uid: zod.string(),
  status: zod.enum(["present", "duplicate", "invalid"]),
  lectureNumber: zod.number(),
  timestamp: zod.coerce.date(),
});
export const GetAttendanceResponse = zod.array(GetAttendanceResponseItem);

/**
 * Records a new RFID attendance scan under the current active lecture
 * @summary Record an attendance scan
 */
export const CreateAttendanceBody = zod.object({
  name: zod.string(),
  uid: zod.string(),
  lectureNumber: zod.number(),
});

/**
 * Returns per-student attendance summary across all lectures
 * @summary Get attendance day summary
 */
export const GetAttendanceSummaryResponseItem = zod.object({
  name: zod.string(),
  uid: zod.string(),
  totalScans: zod.number(),
  presentCount: zod.number(),
  duplicateCount: zod.number(),
  invalidCount: zod.number(),
});
export const GetAttendanceSummaryResponse = zod.array(
  GetAttendanceSummaryResponseItem,
);

/**
 * Returns count breakdown (present, duplicate, invalid) per lecture
 * @summary Get lecture-wise attendance statistics
 */
export const GetLectureStatsResponseItem = zod.object({
  lectureNumber: zod.number(),
  totalScans: zod.number(),
  presentCount: zod.number(),
  duplicateCount: zod.number(),
  invalidCount: zod.number(),
});
export const GetLectureStatsResponse = zod.array(GetLectureStatsResponseItem);

/**
 * @summary Get the currently active lecture
 */
export const GetActiveLectureResponse = zod.object({
  lectureNumber: zod.number().nullable(),
  isActive: zod.boolean(),
  startedAt: zod.coerce.date().nullable(),
});

/**
 * @summary Start a lecture
 */
export const StartLectureBody = zod.object({
  lectureNumber: zod.number(),
});

export const StartLectureResponse = zod.object({
  lectureNumber: zod.number().nullable(),
  isActive: zod.boolean(),
  startedAt: zod.coerce.date().nullable(),
});

/**
 * @summary End the currently active lecture
 */
export const EndLectureResponse = zod.object({
  lectureNumber: zod.number().nullable(),
  isActive: zod.boolean(),
  startedAt: zod.coerce.date().nullable(),
});

/**
 * Clears all attendance records and resets to Lecture 1
 * @summary Reset the entire day
 */
export const ResetDayResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});