# replitui/lib/api-zod/src/generated/types/activeLecture.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Attendance System API
 * OpenAPI spec version: 0.1.0
 */

export interface ActiveLecture {
  /** @nullable */
  lectureNumber: number | null;
  isActive: boolean;
  /** @nullable */
  startedAt: Date | null;
}