# replitui/lib/api-zod/src/generated/types/attendanceRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Attendance System API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceRecordStatus } from "./attendanceRecordStatus";

export interface AttendanceRecord {
  id: number;
  name: string;
  uid: string;
  status: AttendanceRecordStatus;
  lectureNumber: number;
  timestamp: Date;
}