# replitui/lib/api-zod/src/generated/types/createAttendanceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Attendance System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAttendanceBody {
  name: string;
  uid: string;
  lectureNumber: number;
}