# replitui/lib/api-zod/src/generated/types/resetResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Attendance System API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetResponse {
  success: boolean;
  message: string;
}