# replitui/lib/api-zod/src/generated/types/studentSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RFID Attendance System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentSummary {
  name: string;
  uid: string;
  totalScans: number;
  presentCount: number;
  duplicateCount: number;
  invalidCount: number;
}